backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
mediapipe==0.10.18
opencv-python-headless==4.10.0.84
boto3==1.35.0
python-dotenv==1.0.1
pydantic==2.9.0
numpy==1.26.4

sqlalchemy==2.0.36
psycopg2-binary==2.9.9
